+++ axelor-stock/src/main/java/com/axelor/apps/stock/exception/IExceptionMessage.java
@@ -198,7 +198,7 @@ public interface IExceptionMessage {
   String TRACK_NUMBER_WIZARD_TITLE = /*$$(*/ "Enter tracking numbers" /*)*/;
   String TRACK_NUMBER_WIZARD_NO_RECORD_ADDED_ERROR = /*$$(*/ "No Tracking Numbers Added" /*)*/;
 
-  String TRACK_NUMBER_DATE_MISSING = /*$$(*/ "Please filled estimated delivery date" /*)*/;
+  String TRACK_NUMBER_DATE_MISSING = /*$$(*/ "Please fill estimated delivery date for product %s from %s" /*)*/;
 
   /** Stock correction service and controller */
   public static final String STOCK_CORRECTION_1 = /*$$(*/