+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveLineServiceImpl.java
@@ -289,7 +289,7 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
           stockMoveLine,
           minQty,
           trackingNumberService.getTrackingNumber(
-              product, qtyByTracking, stockMove.getCompany(), stockMove.getEstimatedDate()));
+              product, qtyByTracking, stockMove.getCompany(), stockMove.getEstimatedDate(),stockMove.getOrigin()));
 
       generateTrakingNumberCounter++;
 
@@ -303,7 +303,7 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
 
       stockMoveLine.setTrackingNumber(
           trackingNumberService.getTrackingNumber(
-              product, qtyByTracking, stockMove.getCompany(), stockMove.getEstimatedDate()));
+              product, qtyByTracking, stockMove.getCompany(), stockMove.getEstimatedDate(),stockMove.getOrigin()));
     }
   }